% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardioStatsUSA-package.R
\docType{package}
\name{cardioStatsUSA-package}
\alias{cardioStatsUSA}
\alias{cardioStatsUSA-package}
\title{cardioStatsUSA: Analysis and Visualization of Cardiometabolic Outcomes Using NHANES}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analyze and visualize cardiometabolic outcomes among adults living in the United States of America using data from the National Health and Nutrition Examination Survey (NHANES). Save your results, either in the form of a dataset or as a figure. All analyses account for the NHANES design weights and are tested for equivalence with CDC statistics and prior papers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jhs-hwg.github.io/cardioStatsUSA/}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bjaeger@wakehealth.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

Authors:
\itemize{
  \item Ligong Chen \email{ligongc@uab.edu} [contributor]
  \item Paul Muntner \email{pmuntner@uab.edu} [compiler, data contributor]
}

}
\keyword{internal}
