% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_info.R
\name{get_var_info}
\alias{get_var_info}
\alias{get_variable_class}
\alias{get_variable_label}
\alias{get_variable_source}
\alias{get_variable_description}
\alias{get_variable_type}
\alias{get_variable_subpop}
\title{Title}
\usage{
get_var_info(x, key, field)

get_variable_class(x, key)

get_variable_label(x, key)

get_variable_source(x, key)

get_variable_description(x, key)

get_variable_type(x, key)

get_variable_subpop(x, key)
}
\arguments{
\item{field}{}
}
\description{
Title
}
