% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_guide.R
\docType{data}
\name{key_guide}
\alias{key_guide}
\title{Guide to making keys}
\format{
A data frame with 11 rows and 4 variables:
\describe{
\item{name}{The name of the column in the \code{key} data}
\item{type}{The type that this column should be in the \code{key} data}
\item{required}{Is this column required?}
\item{purpose}{What is this part of the \code{key} used for?}
}
}
\usage{
key_guide
}
\description{
A data frame that gives users instructions when they are
making their own keys to extend the NHANES web application.
}
\keyword{datasets}
