% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_bp.R
\docType{data}
\name{nhanes_bp}
\alias{nhanes_bp}
\title{NHANES blood pressure data, 1999-2020}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 63041 rows and 100 columns.
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx}
}
\usage{
nhanes_bp
}
\description{
A dataset containing variables that pertain to blood pressure from the
NHANES continuous examination cycles. The documenation of the variables
in this data are available online (TODO: add link)
}
\keyword{datasets}
