% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_bp_make.R
\name{nhanes_bp_make}
\alias{nhanes_bp_make}
\title{make the data used in shiny app}
\usage{
nhanes_bp_make()
}
\arguments{
\item{write_data}{if \code{TRUE}, the file is written.
If \code{FALSE}, data are returned and no file is written.}
}
\description{
'make' means we can use this function to make a file in the data directory.
}
