% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_design.R
\name{nhanes_design}
\alias{nhanes_design}
\title{Title}
\usage{
nhanes_design(
  data,
  key,
  outcome_variable,
  outcome_quantiles = NULL,
  group_variable = NULL,
  group_cut_n = NULL,
  group_cut_type = NULL,
  stratify_variable = NULL,
  time_variable = "svy_year",
  time_values = NULL,
  pool = FALSE,
  run_checks = TRUE
)
}
\arguments{
\item{run_checks}{}
}
\description{
Title
}
\examples{

x <- nhanes_design(data = nhanes_data,
                   key = nhanes_key,
                   outcome_variable = 'bp_sys_mean')

}
