% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_design_standardize.R
\name{nhanes_design_standardize}
\alias{nhanes_design_standardize}
\title{Title}
\usage{
nhanes_design_standardize(
  x,
  standard_variable = "demo_age_cat",
  standard_weights = c(0.155, 0.454, 0.215, 0.177)
)
}
\arguments{
\item{standard_weights}{}
}
\description{
Title
}
