% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_design_summarize.R
\name{nhanes_design_summarize}
\alias{nhanes_design_summarize}
\title{Summarize an NHANES design}
\usage{
nhanes_design_summarize(x, stats = x$stats, simplify_output = FALSE)
}
\arguments{
\item{x}{[nhanes_design]

an NHANES design object. See \link{nhanes_design} for more details.}

\item{stats}{[character(1+)]

The statistics that should be computed.
Multiple statistics may be requested.
Valid options depend on the type of outcome to be summarized.
For continuous outcomes, valid options include
\itemize{
\item 'mean': estimates the mean value of the outcome
\item 'quantile': estimates 25th, 50th, and 75th percentile of the outcome.
}

For categorical outcomes, valid options include
\itemize{
\item 'percentage': estimates the prevalence of the outcome
\item 'percentage_kg': estimates the prevalence and uses Korn and Graubard's method to estimate a 95\% confidence interval
\item 'count': estimates the number of US adults with the outcome.
}}

\item{simplify_output}{[logical(1)]

The type of output returned will be \link{nhanes_design} if \code{simplify_output}
is \code{FALSE} and a \code{data.table} otherwise.}
}
\value{
if \code{simplify_output} is \code{TRUE}, a \code{data.table}.
Otherwise, a modified \link{nhanes_design} object is returned.
}
\description{
Summarize an NHANES design
}
