% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_design.R
\name{nhanes_design_update}
\alias{nhanes_design_update}
\title{Title}
\usage{
nhanes_design_update(
  x,
  outcome_variable = NULL,
  group_variable = NULL,
  group_cut_n = NULL,
  group_cut_type = NULL,
  stratify_variable = NULL,
  time_values = NULL,
  pool = NULL
)
}
\arguments{
\item{pool}{}
}
\description{
Title
}
