% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_design_viz.R
\name{nhanes_design_viz}
\alias{nhanes_design_viz}
\title{Visualize a summary of NHANES data}
\usage{
nhanes_design_viz(
  x,
  statistic_primary = NULL,
  title = NULL,
  geom = "bar",
  reorder_cats = FALSE,
  width = NULL,
  height = 600,
  size_point = NULL,
  size_error = NULL
)
}
\arguments{
\item{x}{[nhanes_design]

an NHANES design object. See \link{nhanes_design} for more details.}

\item{statistic_primary}{(character) the statistic that defines the
geometric objects in the plot.}

\item{geom}{(character) can be 'bar' or 'points'}

\item{reorder_cats}{(logical) whether to re-order the categorical group
variable so that its levels are shown in increasing order by the expected
outcome.}
}
\value{
a \code{plotly} object
}
\description{
Analyze the prevalence, mean, or quantiles of an outcome
over time.
}
