% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_key.R
\docType{data}
\name{nhanes_key}
\alias{nhanes_key}
\title{NHANES Variables: Names, Roles, and Descriptions}
\format{
A data frame with 97 rows and 11 variables:
\describe{
\item{class}{The broad class of variables that the current variable belongs to}
\item{variable}{The name of the variable in \code{nhanes_data}}
\item{label}{A label that is used to describe the variable in tables and figures}
\item{source}{Which NHANES form did this variable come from?}
\item{type}{What type of variable this is? \code{ctns} indicates continous; \code{catg} categorical; \code{intg} integer, and \code{bnry} binary}
\item{outcome}{\code{TRUE/FALSE} - is this variable allowed to be an outcome?}
\item{group}{\code{TRUE/FALSE} - is this variable allowed to be used for groups?}
\item{subset}{\code{TRUE/FALSE} - is this variable allowed to be used for subsetting?}
\item{stratify}{\code{TRUE/FALSE} - is this variable allowed to be used for stratification?}
\item{module}{What module does this variable belong to? This is used to subset NHANES data to participants who are included in the given module.}
\item{description}{A detailed summary of the variable.}
}
}
\usage{
nhanes_key
}
\description{
A data frame that describes the elements in \code{nhanes_data}. Each row
in \code{nhanes_key} corresponds to a particular variable in \code{nhanes_data},
summarizing that variable's information and what roles the variable
can fulfill in the Shiny application.
}
\details{
Not all variables in \code{nhanes_data} are described in \code{nhanes_key}.
For example, \code{demo_race_black} is not described in \code{nhanes_key} because
it is only used for calibration of survey weights when counts are
estimated (i.e., it has no direct use in the shiny application).
}
\keyword{datasets}
