% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_recode.R
\name{nhanes_recode}
\alias{nhanes_recode}
\title{recode categories of discrete variables in nhanes

instructions are taken from docs/instructions_NHANES.docx}
\usage{
nhanes_recode(data = NULL)
}
\arguments{
\item{data}{if NULL, data are loaded with \code{nhanes_load}}
}
\description{
recode categories of discrete variables in nhanes

instructions are taken from docs/instructions_NHANES.docx
}
