% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesShinyBP-package.R
\docType{package}
\name{nhanesShinyBP-package}
\alias{nhanesShinyBP}
\alias{nhanesShinyBP-package}
\title{nhanesShinyBP: Explore 20+ years of blood pressure data from NHANES}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analyze cardiometabolic outcomes among adults living in the United States of America using data from the National Health and Nutrition Examination Survey (NHANES). Save your results, either in the form of a dataset or as a figure. All analyses account for the NHANES design weights and are tested for equivalence with CDC statistics and prior papers.
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bjaeger@wakehealth.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

Authors:
\itemize{
  \item Paul Muntner \email{pmuntner@uab.edu} [compiler, data contributor]
}

Other contributors:
\itemize{
  \item Ligong Chen \email{ligongc@uab.edu} [contributor]
}

}
\keyword{internal}
