% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_viz.R
\name{plotly_viz}
\alias{plotly_viz}
\title{Visualize a summary of NHANES data}
\usage{
plotly_viz(
  data,
  outcome,
  exposure = NULL,
  group = NULL,
  statistic_primary,
  geom = "bar",
  years,
  pool = "no",
  reorder_cats = FALSE,
  width = NULL,
  height = 600
)
}
\arguments{
\item{data}{a summary data frame produced by \code{svy_design_summarize}}

\item{outcome}{(character) the outcome}

\item{exposure}{(character; optional) the exposure variable}

\item{group}{(character; optional)}

\item{statistic_primary}{(character) the statistic that defines the
geometric objects in the plot.}

\item{geom}{(character) can be 'bar' or 'points'}

\item{years}{(character) the years to be displayed in the plot}

\item{pool}{(character) 'yes' or 'no', is this a trend plot or a
pooled analysis?}

\item{reorder_cats}{(logical) whether to re-order the categorical exposure
variable so that its levels are shown in increasing order by the expected
outcome.}

\item{outcome_type}{(character) valid options are 'ctns', 'bnry', and 'catg'}

\item{stat_all}{(character) a vector containing all the statistics that
are to be included in the plot The primary statistic determines the
geometric objects in the plot. All other statistics are presented in
hover boxes that appear on mouse-over for the relevant object.}
}
\value{
a \code{plotly} object
}
\description{
Analyze the prevalence, mean, or quantiles of an outcome
over time.
}
