% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_tables.R
\docType{data}
\name{replication}
\alias{replication}
\alias{muntner_jama_2020_table_2}
\alias{muntner_jama_2020_etable_1,}
\alias{muntner_hypertension_2022_table_s2,}
\alias{cdc_db289_figure_2,}
\alias{cdc_db289_figure_2}
\alias{muntner_jama_2020_etable_1}
\alias{muntner_hypertension_2022_table_s2}
\alias{cdc_db289_figure_1}
\title{Results from Prior Studies of NHANES Data}
\format{
Data frames with varying numbers of rows and variables.
\describe{
\item{svy_year}{NHANES cycle.}
\item{variable}{variable name.}
\item{group}{sub-group defined by variable.}
\item{estimate}{point estimate.}
\item{std_error}{standard error of the point estimate.}
\item{ci_lower}{lower bound of 95\% confidence interval for the point estimate.}
\item{ci_upper}{upper bound of 95\% confidence interval for the point estimate.}
}
}
\usage{
muntner_jama_2020_table_2

muntner_jama_2020_etable_1

muntner_hypertension_2022_table_s2

cdc_db289_figure_1

cdc_db289_figure_2
}
\description{
These datasets contain values from various publications that have been used
to validate the \code{cardioStatsUSA} shiny application (see \link{app_run}).
}
\keyword{datasets}
