% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_design_new.R
\name{svy_design_new}
\alias{svy_design_new}
\title{Make a new design object}
\usage{
svy_design_new(
  data,
  exposure,
  n_exposure_group,
  exposure_cut_type,
  years,
  pool
)
}
\arguments{
\item{data}{nhanes data - should be recoded (see \link{nhanes_recode})}

\item{exposure}{the exposure variable}

\item{n_exposure_group}{how many exposure groups? only applies to
continuous exposure variables}

\item{exposure_cut_type}{how to cut the exposure groups? only applies to
continuous exposure variables.}

\item{years}{the years to be included in the analysis}

\item{pool}{are the results pooled across multiple NHANES cycles?
if so, say 'yes', o.w. say 'no'}
}
\value{
a \code{svydesign} object
}
\description{
Make a new design object
}
