% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_design_subset.R
\name{svy_design_subset}
\alias{svy_design_subset}
\title{subset a svy design object}
\usage{
svy_design_subset(design, subset_calls)
}
\arguments{
\item{design}{A survey design object.}

\item{subset_var}{(\emph{character}\link{length 1}) the variable to subset with.}

\item{subset_values}{(\emph{character}\link{length 1+}) the values of the subset
variable that will be retained.}
}
\description{
subset a svy design object
}
