% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_design_summarize.R
\name{svy_design_summarize}
\alias{svy_design_summarize}
\title{Summarize a survey design}
\usage{
svy_design_summarize(
  design,
  outcome,
  statistic = NULL,
  exposure = NULL,
  group = NULL,
  pool_svy_years = FALSE,
  simplify_bnry_output = TRUE,
  quantiles = c(0.25, 0.5, 0.75),
  age_standardize = FALSE,
  age_wts = c(0.155, 0.454, 0.215, 0.177)
)
}
\arguments{
\item{design}{an object created by \link{svy_design_new}}

\item{outcome}{(\emph{character}) the outcome variable}

\item{statistic}{(\emph{character vector}) which statistics to compute}

\item{exposure}{(\emph{character}) the exposure variable}

\item{group}{(\emph{character}) the group variable}

\item{pool_svy_years}{(\emph{logical}) \code{TRUE} in pool is 'yes', \code{FALSE} o.w.}

\item{simplify_bnry_output}{not currently used}

\item{quantiles}{(\emph{numeric vector}) percentiles for stat summaries}

\item{age_standardize}{(\emph{logical}) standardize by age groups?}

\item{age_wts}{(\emph{numeric vector}) weights for age standardization.}

\item{key}{should always be \code{nhanes_key}}
}
\value{
summaries
}
\description{
Summarize a survey design
}
