% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_stat_tidy.R
\name{svy_stat_tidy}
\alias{svy_stat_tidy}
\title{standardize survey function outputs}
\usage{
svy_stat_tidy(x, outcome, by_vars)
}
\arguments{
\item{x}{output from one of the stat functions.}

\item{outcome}{the outcome variable}

\item{by_vars}{variables to stratify by}
}
\value{
a data frame
}
\description{
standardize survey function outputs
}
