% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{app_run}
\alias{app_run}
\title{Run the NHANES BP application}
\usage{
app_run(...)
}
\arguments{
\item{...}{currently not used}
}
\value{
runs a shiny application locally
}
\description{
Run the NHANES BP application
}
