% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_info.R
\name{get_var_info}
\alias{get_var_info}
\alias{get_variable_class}
\alias{get_variable_label}
\alias{get_variable_source}
\alias{get_variable_description}
\alias{get_variable_type}
\alias{get_variable_module}
\title{Get variable information}
\usage{
get_var_info(x, key, field)

get_variable_class(x, key)

get_variable_label(x, key)

get_variable_source(x, key)

get_variable_description(x, key)

get_variable_type(x, key)

get_variable_module(x, key)
}
\arguments{
\item{x}{[character(1)]

a variable name that is present in the \code{variable} column of \code{key}}

\item{key}{[data.frame]

A data set with one row per variable and with columns that describe the variable. See \link{nhanes_key} for more details}

\item{field}{[character(1)]

the column name in \code{nhanes_key} to get information from.}
}
\value{
a character value
}
\description{
Get variable information
}
\examples{

# the low-level working function
get_var_info('bp_sys_mean', key = nhanes_key, field = 'class')

# the convenience functions
get_variable_class('bp_sys_mean')
get_variable_label('bp_sys_mean')
get_variable_source('bp_sys_mean')
get_variable_description('bp_sys_mean')
get_variable_module('bp_sys_mean')

}
