% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_data.R
\docType{data}
\name{nhanes_data}
\alias{nhanes_data}
\title{NHANES data, 1999-2020}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 63041 rows and 98 columns.
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx}
}
\usage{
nhanes_data
}
\description{
The NHANES program was initiated in the early 1960s and beginning in 1999 has been conducted continuously, in two-year cycles. The protocols for each cycle were approved by the NCHS Institutional Review Board. Written informed consent was obtained from each participant.
}
\details{
NHANES data include in-home interviews and study examinations conducted at mobile examination centers. The interview included questions about demographics, health behaviors, prior diagnoses, medication use, and medical history. During the interview, the labels of medications that participants reported taking in the preceding 30 days were recorded. During the study examination, height and weight were measured, blood samples were used to measure cholesterol, glycated hemoglobin and serum creatinine, and a spot urine sample was used to measure albumin and creatinine and to conduct a pregnancy test.
}
\section{Survey variables}{
In each NHANES cycle, potential participants were identified using a
multi-stage sampling process. The variables below are based on this
process.
\subsection{svy_id}{
\itemize{
\item Label: participant identifier
\item Description: NHANES participant unique identifier.
}
}

\subsection{svy_psu}{
\itemize{
\item Label: primary sampling unit
\item Description: Population sampling unit. This variable is used to
account for the non-random selection of study participants for
NHANES
}
}

\subsection{svy_strata}{
\itemize{
\item Label: strata
\item Description: Population stratification. This variable is used to
account for the non-random selection of study participants for
NHANES
}
}

\subsection{svy_weight_mec}{
\itemize{
\item Label: Mobile examination center weights
\item Description: Weight applied to produce statistical estimates for the
non-institutionalized US population. This weight is used for
calculating means and proportions.
}
}

\subsection{svy_subpop_htn}{
\itemize{
\item Label: Subpopulation for hypertension
\item Description: This indicates that the person has data needed to be
included in the analysis of blood pressure or hypertension data
(i.e., they had at least one systolic and diastolic blood pressure
measurement and they had information on self-reported
antihypertensive medication use).
}
}

\subsection{svy_year}{
\itemize{
\item Label: NHANES cycle
\item Description: NHANES survey cycle: 1999-2000, 2001-2002, 2003-2004,
2005-2006, 2007-2008, 2009-2010, 2011-2012, 2013-2014, 2015-2016,
2017-2020
}
}
}

\section{Demographics variables}{
The NHANES in-home interview included questions about participant
demographics, including age, sex, race/ethnicity, and pregnancy status
\subsection{demo_age_cat}{
\itemize{
\item Label: Age category, years
\item Description: Age grouping: 18-44, 45-64, 65-74, >= 75 years
}
}

\subsection{demo_race}{
\itemize{
\item Label: Race
\item Description: Self-reported race/ethnicity. From 1999-2000 through
2009-2010 this was available as non-Hispanic White, non-Hispanic
Black, Hispanic and other. From 2011-2012 through 2017-2020 this was
available as non-Hispanic White, non-Hispanic Black, non-Hispanic
Asian, Hispanic and other.
}
}

\subsection{demo_age_years}{
\itemize{
\item Label: Age, years
\item Description: Participant age in years. Participants > 80 years of
age are given an age of 80 years.
}
}

\subsection{demo_pregnant}{
\itemize{
\item Label: Pregnant
\item Description: Pregnancy status. This is defined by either self-report
of being pregnant or a positive pregnancy test conducted during the
study visit.
}
}

\subsection{demo_gender}{
\itemize{
\item Label: Gender
\item Description: Self-reported gender
}
}
}

\section{Blood pressure variables}{
Systolic and diastolic BP (SBP and DBP, respectively) were measured
three times by trained and certified physicians. BP was measured using a
mercury sphygmomanometer from 1999-2000 through 2015-2016 and using an
oscillometric device in 2017-2020. The oscillometric SBP and DBP values
were calibrated to the mercury device. Mean SBP and DBP levels were
computed over all available
\subsection{bp_sys_mean}{
\itemize{
\item Label: Systolic blood pressure, mm Hg
\item Description: Mean systolic blood pressure in mm Hg. This is based on
the average of up to 3 readings. Overall, >95\% of participants with
at least one systolic blood pressure reading had three readings.
From 1999-2000 through 2015-2016, systolic blood pressure was
measured using a mercury sphygmomanometer. In 2017-2020, systolic
blood pressure was measured using an oscillometric device. The
systolic blood pressure in 2017-2020 was calibrated to the mercury
device by adding 1.5 mm Hg to the mean measured value.
}
}

\subsection{bp_dia_mean}{
\itemize{
\item Label: Diastolic blood pressure, mm Hg
\item Description: Mean diastolic blood pressure in mm Hg. This is based
on the average of up to 3 readings. Overall, >95\% of participants
with at least one diastolic blood pressure reading had three
readings. From 1999-2000 through 2015-2016, diastolic blood pressure
was measured using a mercury sphygmomanometer. In 2017-2020,
diastolic blood pressure was measured using an oscillometric device.
The diastolic blood pressure in 2017-2020 was calibrated to the
mercury device by subtracting 1.0 mm Hg to the mean measured value.
}
}

\subsection{bp_cat_meds_excluded}{
\itemize{
\item Label: Blood pressure category
\item Description: Systolic/diastolic blood pressure <120/80,
120-129/<80, 130-139/80-89, 140-159/90-99, >= 160/100.
Participants were placed in the category associated with higher
blood pressure (e.g., someone with systolic blood pressure of 150 mm
Hg and diastolic blood pressure of 76 mm Hg was placed in the >=
140/90 mm Hg category)
}
}

\subsection{bp_cat_meds_included}{
\itemize{
\item Label: Blood pressure category (including antihypertensive
medication use)
\item Description: Systolic/diastolic blood pressure <120/80,
120-129/<80, 130-139/80-89, 140-159/90-99, >= 160/100.
Participants taking antihypertensive medication were placed in a
separate category. Participants were placed in the category
associated with higher blood pressure (e.g., someone with systolic
blood pressure of 150 mm Hg and diastolic blood pressure of 76 mm Hg
was placed in the >= 140/90 mm Hg category)
}
}

\subsection{bp_control_jnc7}{
\itemize{
\item Label: Blood pressure control (JNC7 guideline)
\item Description: Systolic and diastolic blood pressure controlled to the
levels recommended in the JNC7 guideline, systolic blood pressure <
140 mm Hg and diastolic blood pressure < 90 mm Hg.
}
}

\subsection{bp_control_accaha}{
\itemize{
\item Label: Blood pressure control (2017 ACC/AHA BP guideline)
\item Description: Systolic and diastolic blood pressure controlled to the
levels recommended in the 2017 ACC/AHA BP guideline, systolic blood
pressure < 130 mm Hg and diastolic blood pressure < 80 mm Hg
except for those >= 65 years of age without diabetes, chronic
kidney disease, history of cardiovascular disease or 10-year
predicted ASCVD risk >= 10\%. For this group, blood pressure control
was defined as systolic blood pressure < 130 mm Hg
}
}

\subsection{bp_uncontrolled_jnc7}{
\itemize{
\item Label: Uncontrolled BP (JNC7 guideline)
\item Description: Systolic blood pressure >= 140 mm Hg or diastolic
blood pressure >= 90 mm Hg
}
}

\subsection{bp_uncontrolled_accaha}{
\itemize{
\item Label: Uncontrolled BP (2017 ACC/AHA BP guideline)
\item Description: Systolic blood pressure >= 130 mm Hg or diastolic
blood pressure >= 80 mm Hg except for individuals without diabetes,
chronic kidney disease, history of cardiovascular disease or 10-year
predicted ASCVD risk >= 10\%. For this group, uncontrolled blood
pressure was defined as systolic blood pressure >= 130 mm Hg
}
}
}

\section{Hypertension variables}{
These variables are derived according to the 2017 American College of
Cardiology/American Heart Association (ACC/AHA) BP guideline and the
Seventh Joint National Committee (JNC7) BP guideline.
\subsection{htn_jnc7}{
\itemize{
\item Label: Hypertension (JNC7 guideline)
\item Description: Hypertension defined by the JNC7 guideline, systolic
blood pressure >= 140 mm Hg, diastolic blood pressure >= 90 mm Hg
or self-reported antihypertensive medication use.
}
}

\subsection{htn_accaha}{
\itemize{
\item Label: Hypertension (2017 ACC/AHA BP guideline)
\item Description: Hypertension defined by the 2017 ACC/AHA blood pressure
guideline, systolic blood pressure >= 130 mm Hg, diastolic blood
pressure >= 80 mm Hg or self-reported antihypertensive medication
use.
}
}

\subsection{htn_aware}{
\itemize{
\item Label: Awareness of hypertension
\item Description: Self-report of a prior diagnosis of antihypertensive
medication.
}
}

\subsection{htn_resistant_jnc7}{
\itemize{
\item Label: Resistant hypertension (JNC7 guideline definition)
\item Description: Taking 4 or more classes of antihypertensive
medication, systolic blood pressure >= 140 mm Hg or diastolic blood
pressure >= 90 mm Hg with the use of 3 classes of antihypertensive
medication, or systolic blood pressure >= 130 mm Hg or diastolic
blood pressure >= 80 mm Hg with the use of 3 classes of
antihypertensive medication for those with diabetes or chronic
kidney disease.
}
}

\subsection{htn_resistant_accaha}{
\itemize{
\item Label: Resistant hypertension (2017 ACC/AHA BP guideline)
\item Description: Taking 4 or more classes of antihypertensive
medication, systolic blood pressure >= 140 mm Hg or diastolic blood
pressure >= 90 mm Hg with the use of 3 classes of antihypertensive
medication, or systolic blood pressure >= 130 mm Hg or diastolic
blood pressure >= 80 mm Hg with the use of 3 classes of
antihypertensive medication for those <65 years of age and those
>= 65 years of age with diabetes, chronic kidney disease or high
cardiovascular risk defined by a history of cardiovascular disease
or 10-year predicted risk >= 10\% using the pooled cohort risk
equations.
}
}

\subsection{htn_resistant_accaha_thz}{
\itemize{
\item Label: Resistant hypertension (JNC7 guideline, requires thiazide
diuretic)
\item Description: Taking 4 or more classes of antihypertensive
medication, systolic blood pressure >= 140 mm Hg or diastolic blood
pressure >= 90 mm Hg with the use of 3 classes of antihypertensive
medication, or systolic blood pressure >= 130 mm Hg or diastolic
blood pressure >= 80 mm Hg with the use of 3 classes of
antihypertensive medication for those < 65 years of age and those
>= 65 years of age with diabetes, chronic kidney disease or high
cardiovascular risk defined by a history of cardiovascular disease
or 10-year predicted risk >= 10\% using the pooled cohort risk
equations. To meet this definition of resistant hypertension, the
participant had to be taking a thiazide diuretic.
}
}

\subsection{htn_resistant_jnc7_thz}{
\itemize{
\item Label: Resistant hypertension (2017 ACC/AHA BP guideline, requires
thiazide diuretic)
\item Description: Taking 4 or more classes of antihypertensive
medication, systolic blood pressure >= 140 mm Hg or diastolic blood
pressure >= 90 mm Hg with the use of 3 classes of antihypertensive
medication, or systolic blood pressure >= 130 mm Hg or diastolic
blood pressure >= 80 mm Hg with the use of 3 classes of
antihypertensive medication for those with diabetes or chronic
kidney disease. To meet this definition of resistant hypertension,
the participant had to be taking a thiazide diuretic.
}
}
}

\section{Antihypertensive medication variables}{
During the interview, the labels of medications that participants
reported taking in the preceding 30 days were recorded.
\subsection{bp_med_use}{
\itemize{
\item Label: Self-reported antihypertensive medication use
\item Description: Self-reported use of antihypertensive medication
}
}

\subsection{bp_med_recommended_jnc7}{
\itemize{
\item Label: Antihypertensive medications recommended (JNC7 guideline)
\item Description: Systolic blood pressure >= 140 mm Hg or diastolic
blood pressure >= 90 mm Hg; Systolic blood pressure >= 130 mm Hg
or diastolic blood pressure >= 80 mm Hg for those with chronic
kidney disease or diabetes. Those taking antihypertensive
medications were considered to be recommended treatment by this
definition.
}
}

\subsection{bp_med_recommended_accaha}{
\itemize{
\item Label: Antihypertensive medications recommended (2017 ACC/AHA BP
guideline)
\item Description: Systolic blood pressure >= 140 mm Hg or diastolic
blood pressure >= 90 mm Hg; Systolic blood pressure >= 130 mm Hg
or diastolic blood pressure >= 80 mm Hg for those with chronic
kidney disease, diabetes, 10-year predicted atherosclerotic
cardiovascular disease risk by the pooled cohort risk equations or
age >= 65 years. Those taking antihypertensive medications were
considered to be recommended treatment by this definition.
}
}

\subsection{bp_med_n_class}{
\itemize{
\item Label: Number of antihypertensive medication classes
\item Description: Number of antihypertensive medication classes being
taken based on the pill bottle review
}
}

\subsection{bp_med_ace}{
\itemize{
\item Label: ACE inhibitors
\item Description: Taking an angiotensin converting enzyme inhibitor,
defined using the pill bottle review. Drugs in this class included
bnazepril, captopril, enalapril, fosinopril, lisonopril, moexipril,
perindopril, quinapril, ramipril, trandolapril
}
}

\subsection{bp_med_aldo}{
\itemize{
\item Label: Aldosterone antagonists
\item Description: Taking an aldosterone antagonist, defined using the
pill bottle review. Drugs in this class included eplerenone,
spironolactone.
}
}

\subsection{bp_med_alpha}{
\itemize{
\item Label: Alpha-1 blockers
\item Description: Taking an alpha blocker, defined using the pill bottle
review. Drugs in this class included doxazosin, prazosin, terazosin.
}
}

\subsection{bp_med_angioten}{
\itemize{
\item Label: Angiotensin receptor blockers
\item Description: Taking an angiotensin receptor blocker, defined using
the pill bottle review. Drugs in this class included candesartan,
eprosartan, irbesartan, losartan, olmesartan, telmisartan,
valsartan, azilsartan.
}
}

\subsection{bp_med_beta}{
\itemize{
\item Label: Beta blockers
\item Description: Taking a beta blocker. Drugs in this class included
acebutolol, atenolol, betaxolol, bisoprolol, carvedilol, labetalol,
metoprolol, nadolol, nebivolol, pindolol, propranolol.
}
}

\subsection{bp_med_central}{
\itemize{
\item Label: Central alpha1 agonist and other centrally acting agents
\item Description: Taking a centrally acting agents, defined using the
pill bottle review. Drugs in this class included clonidine,
guanabenz, guanfaacine, methyldopa, reserpine.
}
}

\subsection{bp_med_ccb}{
\itemize{
\item Label: Calcium channel blockers
\item Description: Taking a calcium channel blocker, defined using the
pill bottle review. Drugs in this class included amlodipine,
diltiazem, felodipine, isradipine, nicardipine, nifedipine,
nisoldipine, verapamil.
}
}

\subsection{bp_med_diur_Ksparing}{
\itemize{
\item Label: Potassium sparing diuretics
\item Description: Taking a potassium-sparing diuretic, defined using the
pill bottle review. Drugs in this class included amiloride,
triamterene.
}
}

\subsection{bp_med_diur_loop}{
\itemize{
\item Label: Loop diuretics
\item Description: Taking a loop diuretic, defined using the pill bottle
review. Drugs in this class included bumetanide, furosemide,
torsemide, ethacrynic acid.
}
}

\subsection{bp_med_diur_thz}{
\itemize{
\item Label: Thiazide or thiazide-type diuretics
\item Description: Taking a thiazide diuretic, defined using the pill
bottle review. Drugs in this class included bendroflumethiazide,
chlorthalidone, chlorothiazide, hydrochlorothiazide, indapamide,
metolazone, polythiazide.
}
}

\subsection{bp_med_renin_inhibitors}{
\itemize{
\item Label: Direct renin inhibitors
\item Description: Taking a renin inhibitor, defined using the pill bottle
review. Drugs in this class included aliskiren.
}
}

\subsection{bp_med_vasod}{
\itemize{
\item Label: Direct vasodilators
\item Description: Taking a vasodilator, defined using the pill bottle
review. Drugs in this class included hydralazine, minoxidil.
}
}
}

\section{Comorbidities variables}{
These variables describe comorbid conditions.
\subsection{cc_n_highrisk}{
\itemize{
\item Label: Number of high risk conditions
\item Description: Self-reported history of coronary heart disease,
myocardial infarction, stroke or heart failure or 10-year predicted
risk >= 10\% estimated by the pooled cohort risk equations
}
}

\subsection{cc_smoke}{
\itemize{
\item Label: Smoking status
\item Description: Self-reported current cigarette smoking
}
}

\subsection{cc_bmi}{
\itemize{
\item Label: Body mass index, kg/m2
\item Description: Body mass index in kg/m2, estimated using the height
and weight measured during the study examination.
}
}

\subsection{cc_diabetes}{
\itemize{
\item Label: Prevalent diabetes
\item Description: HbA1c >= 6.5\% or self-report of a prior diagnosis of
diabetes with use of insulin or oral glucose-lowering medications.
}
}

\subsection{cc_ckd}{
\itemize{
\item Label: Prevalent chronic kidney disease
\item Description: Estimated glomerular filtration rate < 60 ml/min/1.73
m2 or albumin-to-creatinine > 30 mg/g. Estimated glomerular
filtration rate was calculated using the 2021 serum creatinine-based
equation.
}
}

\subsection{cc_cvd_mi}{
\itemize{
\item Label: History of myocardial infarction
\item Description: Self-reported history of myocardial infarction
}
}

\subsection{cc_cvd_chd}{
\itemize{
\item Label: History of coronary heart disease
\item Description: Self-reported history of myocardial infarction or
coronary heart disease
}
}

\subsection{cc_cvd_stroke}{
\itemize{
\item Label: History of stroke
\item Description: Self-reported history of stroke
}
}

\subsection{cc_cvd_ascvd}{
\itemize{
\item Label: History of ASCVD
\item Description: Self-reported history of coronary heart disease,
myocardial infarction or stroke
}
}

\subsection{cc_cvd_hf}{
\itemize{
\item Label: History of heart failure
\item Description: Self-reported history of heart failure
}
}

\subsection{cc_cvd_any}{
\itemize{
\item Label: History of CVD
\item Description: Self-reported history of coronary heart disease,
myocardial infarction, stroke or heart failure
}
}
}

\keyword{datasets}
