% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_design_standardize.R
\name{nhanes_design_standardize}
\alias{nhanes_design_standardize}
\title{Standardize an NHANES design}
\usage{
nhanes_design_standardize(
  x,
  standard_variable = "demo_age_cat",
  standard_weights = c(0.155, 0.454, 0.215, 0.177)
)
}
\arguments{
\item{x}{[nhanes_design]

an NHANES design object. See \link{nhanes_design} for more details.}

\item{standard_variable}{[character(1)]

The name of the variable used to create standardization groups.
The default is to use \code{demo_age_cat}, which leads to age
standardization.}

\item{standard_weights}{[numeric(n)]

The proportionate weights for each group defined by the standard variable.
The number of weights should equal the number of groups defined by
\code{standard_variable} and all weights must be >0.}
}
\value{
an \link{nhanes_design} object.
}
\description{
Standardize an NHANES design
}
