% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_key.R
\docType{data}
\name{nhanes_key}
\alias{nhanes_key}
\title{One-stop shop for variable names, roles, labels, and functions}
\format{
A data frame with 59799 rows and 35 variables:
\describe{
\item{price}{price, in US dollars}
\item{data}{fill in}
\item{variables}{fill in}
\item{recoder}{fill in}
\item{fctrs}{fill in}
\item{svy_funs}{fill in}
\item{svy_calls}{fill in}
\item{time_var}{fill in}

}
}
\usage{
nhanes_key
}
\description{
A list with various elements pertaining the \code{nhanes_data}
}
\keyword{datasets}
