% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_summarize.R
\name{nhanes_summarize}
\alias{nhanes_summarize}
\title{Generate summaries of NHANES}
\usage{
nhanes_summarize(
  data,
  key,
  outcome_variable,
  outcome_quantiles = NULL,
  group_variable = NULL,
  group_cut_n = NULL,
  group_cut_type = NULL,
  stratify_variable = NULL,
  time_variable = "svy_year",
  time_values = NULL,
  pool = FALSE,
  subset_calls = list(),
  age_wts = NULL,
  simplify_output = TRUE
)
}
\arguments{
\item{pool}{(\emph{logical}; length 1; default = \code{FALSE})}

\item{age_wts}{(\emph{numeric}; length 4; default is none)}

\item{outcome}{(\emph{character}; length 1; no default)
The variable to be summarized.}

\item{exposure}{(\emph{character}; length 1; default is none)}

\item{group}{(\emph{character}; length 1; default is none)}

\item{years}{(\emph{character}; length may vary; default = 'all')
Valid options are
\itemize{
\item 'all' (includes all NHANES cycles)
\item 'last_5' (includes the last 5 most recent NHANES cycles)
\item 'most_recent' (include the most recent NHANES cycle)
}}

\item{subset_variables}{(\emph{list}; length may vary)}

\item{subset_values}{(\emph{list}; length must match \code{subset_variables})}
}
\value{
a \code{data.table} with summarized values
}
\description{
The description should include
\itemize{
\item what is NHANES
\item Who is included in the data to be summarized
\item How is the summary computed
}
}
\examples{

nhanes_summarize("bp_sys_mean")

}
