% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_visualize.R
\name{nhanes_visualize}
\alias{nhanes_visualize}
\title{Visualize summaries of NHANES}
\usage{
nhanes_visualize(
  outcome,
  exposure = NULL,
  group = NULL,
  years = "last_5",
  pool = FALSE,
  subset_variables = list(),
  subset_values = list(),
  age_wts = NULL,
  statistic_primary = NULL,
  geom = "bar",
  reorder_cats = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{outcome}{(\emph{character}; length 1; no default)
The variable to be summarized.}

\item{exposure}{(\emph{character}; length 1; default is none)}

\item{group}{(\emph{character}; length 1; default is none)}

\item{years}{(\emph{character}; length may vary; default = 'all')
Valid options are
\itemize{
\item 'all' (includes all NHANES cycles)
\item 'last_5' (includes the last 5 most recent NHANES cycles)
\item 'most_recent' (include the most recent NHANES cycle)
}}

\item{pool}{(\emph{logical}; length 1; default = \code{FALSE})}

\item{subset_variables}{(\emph{list}; length may vary)}

\item{subset_values}{(\emph{list}; length must match \code{subset_variables})}

\item{age_wts}{(\emph{numeric}; length 4; default is none)}

\item{statistic_primary}{(character) the statistic that defines the
geometric objects in the plot.}

\item{geom}{(character) can be 'bar' or 'points'}

\item{reorder_cats}{(logical) whether to re-order the categorical exposure
variable so that its levels are shown in increasing order by the expected
outcome.}

\item{width}{width of the plot, in pixels}

\item{height}{height of the plot, in pixels}
}
\value{
a plotly visualization
}
\description{
Visualize summaries of NHANES
}
