% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_visualize.R
\name{nhanes_visualize}
\alias{nhanes_visualize}
\title{Visualize summaries of NHANES}
\usage{
nhanes_visualize(
  data,
  key,
  outcome_variable,
  outcome_quantiles = NULL,
  group_variable = NULL,
  group_cut_n = NULL,
  group_cut_type = NULL,
  stratify_variable = NULL,
  time_variable = "svy_year",
  time_values = NULL,
  pool = FALSE,
  subset_variables = list(),
  subset_values = list(),
  age_wts = NULL,
  statistic_primary = NULL,
  geom = "bar",
  reorder_cats = FALSE,
  width = NULL,
  height = NULL,
  size_point = NULL,
  size_error = NULL
)
}
\arguments{
\item{pool}{(\emph{logical}; length 1; default = \code{FALSE})}

\item{subset_variables}{(\emph{list}; length may vary)}

\item{subset_values}{(\emph{list}; length must match \code{subset_variables})}

\item{age_wts}{(\emph{numeric}; length 4; default is none)}

\item{statistic_primary}{(character) the statistic that defines the
geometric objects in the plot.}

\item{geom}{(character) can be 'bar' or 'points'}

\item{reorder_cats}{(logical) whether to re-order the categorical exposure
variable so that its levels are shown in increasing order by the expected
outcome.}

\item{width}{width of the plot, in pixels}

\item{height}{height of the plot, in pixels}
}
\value{
a plotly visualization
}
\description{
Visualize summaries of NHANES
}
