% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesShinyBP-package.R
\docType{package}
\name{nhanesShinyBP-package}
\alias{nhanesShinyBP}
\alias{nhanesShinyBP-package}
\title{nhanesShinyBP: Explore 20+ years of blood pressure data from NHANES}
\description{
Use a simple interface to generate summary data and figures from the continuous NHANES exam, beginning in 1999-2000 and continuing through 2017-2020. Save your results, either in the form of a dataset or as a figure saved to png format. All analyses account for the NHANES design weights and are tested for equivalence with CDC statistics and reports.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
